+++ packages/sdk/svelte/src/lib/index.ts
@@ -1,5 +1,5 @@
 // Reexport your entry components here
-export * as LDClient from './client/SvelteLDClient.js';
+export { LD } from './client/SvelteLDClient.js';
 
 // Export Components
 export { default as LDProvider } from './provider/LDProvider.svelte';