+++ packages/shared/sdk-server/src/api/index.ts
@@ -4,6 +4,7 @@ export * from './LDClient';
 export * from './LDMigration';
 export * from './interfaces/DataKind';
 export * from './subsystems/LDFeatureStore';
+export * from './subsystems/LDTransactionalFeatureStore';
 export * from './LDWaitForInitializationOptions';
 
 // These are items that should be less frequently used, and therefore they