+++ packages/shared/sdk-server/src/store/index.ts
@@ -1,5 +1,13 @@
 import AsyncStoreFacade from './AsyncStoreFacade';
+import AsyncTransactionalStoreFacade from './AsyncTransactionalStoreFacade';
 import PersistentDataStoreWrapper from './PersistentDataStoreWrapper';
 import { deserializePoll } from './serialization';
+import TransactionalFeatureStore from './TransactionalFeatureStore';
 
-export { AsyncStoreFacade, PersistentDataStoreWrapper, deserializePoll };
+export {
+  AsyncStoreFacade,
+  AsyncTransactionalStoreFacade,
+  PersistentDataStoreWrapper,
+  TransactionalFeatureStore,
+  deserializePoll,
+};